Freeze missing-data dataset
Due to changes in R’s implementation of sample() the dataset generated may differ depending on the version of R and loaded packages. Rather than generating the dataset on the fly, the app now loads a static dataset generated with the intended behavior. The old data generation code is retained unused to allow changes.